added AEXML source directly instead of linking to the framework

+++ DocX/NSAttributedString+DocX.swift
@@ -7,25 +7,34 @@
 //
 
 import Foundation
-import AEXML
 import ZipArchive
 
 extension NSAttributedString:DocX{
         
     @objc(saveToUrl:error:) public func saveTo(url:URL)throws{
         
-        let tempURL=FileManager.default.temporaryDirectory.appendingPathComponent(UUID().uuidString, isDirectory: false)
+        let tempURL=try FileManager.default.url(for: .itemReplacementDirectory, in: .userDomainMask, appropriateFor: url, create: true)
+        
+        defer{
+            try? FileManager.default.removeItem(at: tempURL)
+        }
+        
+        let docURL=tempURL.appendingPathComponent(UUID().uuidString, isDirectory: true)
         let attributes=[NSAttributedString.DocumentAttributeKey.documentType:NSAttributedString.DocumentType.officeOpenXML]
         let wrapper=try self.fileWrapper(from: NSRange(location: 0, length: self.length), documentAttributes: attributes)
-        try wrapper.write(to: tempURL, options: .atomic, originalContentsURL: nil)
-        let docPath=tempURL.appendingPathComponent("word").appendingPathComponent("document").appendingPathExtension("xml")
-        try FileManager.default.removeItem(at: docPath)
-        let xmlData = try self.docXDocument()
-        try xmlData.write(to: docPath, atomically: true, encoding: .utf8)
-        let zipURL=FileManager.default.temporaryDirectory.appendingPathComponent(UUID().uuidString).appendingPathExtension("zip")
-        let success=SSZipArchive.createZipFile(atPath: zipURL.path, withContentsOfDirectory: tempURL.path, keepParentDirectory: false)
+        try wrapper.write(to: docURL, options: .atomic, originalContentsURL: nil)
+        
+        if self.containsRubyAnnotations{ // this is the main attribute that is not conserved by the cocoa exporter. there might, potentially, be others
+            let docPath=docURL.appendingPathComponent("word").appendingPathComponent("document").appendingPathExtension("xml")
+            try FileManager.default.removeItem(at: docPath)
+            let xmlData = try self.docXDocument()
+            try xmlData.write(to: docPath, atomically: true, encoding: .utf8)
+        }
+
+        let zipURL=tempURL.appendingPathComponent(UUID().uuidString).appendingPathExtension("zip")
+        let success=SSZipArchive.createZipFile(atPath: zipURL.path, withContentsOfDirectory: docURL.path, keepParentDirectory: false)
         guard success == true else{throw DocXSavingErrors.compressionFailed}
-        try FileManager.default.removeItem(at: tempURL)
+    
         do{
             try FileManager.default.copyItem(at: zipURL, to: url)
         }
@@ -38,8 +47,6 @@ extension NSAttributedString:DocX{
                 throw error
             }
         }
-        try FileManager.default.removeItem(at: zipURL)
-        
     }
     
 }
